Further refactoring of vprrd.py, and a hack to fix the "black load" syndrome.
The hack is called a timewarp; if the data values retrieved from an RRD are
NaN values and timewarping is enabled, we skew the start/stop times by -60
seconds and try to extract those numbers instead.  This will allow Cricket up
to 60 seconds to process its targets before updating an RRD file, and should
ensure that no unnecessary "black load" appears in the Traffic Map applet.

* vprrd.py
  Allow querying any number of datasources per RRD.  Introduce the concept of
  timewarping.
  (rrdvalues): New function.