Revised startup shell scripts for Java programs.
The main new feature is that JAVA_HOME is only used if found either in the
environment or in nav.conf.  If not found, the scripts attempt to use a java
executable found on the PATH.  This enables users to configure NAV to run
using an alternative JVM installation on the same server.

* doc/conf/nav.conf
  Comment out the JAVA_HOME setting, and update other comments.

* src/INSTALL
* src/Makefile.in
* src/networkDiscovery/Main-Class
  Bugfixes, networkDiscovery was wrongly referred to as networkTopology in
  several places.

* src/eventEngine/eventEngine.sh.in
* src/getBoksMacs/getBoksMacs.sh.in
* src/getDeviceData/getDeviceData.sh.in
* src/networkDiscovery/networkDiscovery.sh.in
  Refactored for the JAVA_HOME change.  Other improvements include using the
  -jar option to the JVM, and declaring NAV's javalibdir as one of the JVM's
  java.ext.dirs.