Use more reasonable timeout policy when checking what SNMP version a device supports.
The existing policy was to allow only _one_ timeout of 500ms was too strict,
sometimes causing gDD to think a device is a SNMPv1 device.  This means we lose
the ability to query things such as 64-bit counters, and sometimes causes the
OID tester to wrongly classify the ifHC* counters as non-compatible with this
device.

* SimpleSnmp/src/no/ntnu/nav/SimpleSnmp/SimpleSnmp.java
  (checkSnmpVersion): Set socket timeout to 1000ms, allow the default number of
  timeouts before giving up.