Enable nav.Snmp to communicate with HP SwitchStack members.
By specifiying the target agent when calling the pysnmp.role.manager
constructor, the datagram socket is connected to this peer (i.e. only
datagrams received from this ip and port are accepted by the socket
library).  This is all fine and dandy when talking to regular SNMP
agents that reply from the same port (161) they were queried on.  When
talking to HP SwitchStack members via the commander, using a modified
community string, the commander will reply from port 1507 instead of
161, causing the replies to be discarded by the socket library.

These changes specify the target's ip address and destination port on
each call to manager.send_and_receive(), making sure that the datagram
socket is not connected to the peer, and will accept any packet from
the network.