1Refactored the javascript code that shows preloaded advanced search field because of a race condition that occured.
The only negative with this refactoring is that the code may not have changed
the href because until the page is fully loaded. That said, I see this as a more
stable way of doing it, as it will always change the link to the intended in the
end, script or no script.