Bugfix for LP#262303 (Searching an entire IPv6 subnet hangs the Machine Tracker)
The machine tracker iterates all IP addresses in the range between FROM and TO,
to support the "show inactive addresses" feature.  This is a bad idea(tm) for
IPv6 addresses.

This patch rewrites the iteration to only iterate over addresses that are to be
part of the output, and will also disable display of inactive addresses for
IPv6.

New variables/functions introduced are named according to PEP-8, old names are
kept.  For extra confusion ;-)

* subsystem/machinetracker/src/machinetracker.py
  (ip_range): New function.
  (IPSQLQuery.getRows): Refactored most of the workings of this method. Always
  set nonActive to False if search space is IPv6.  Pre-generate a sequence of
  IP addresses to output, based on the search parameters, then iterate this
  sequence.

* subsystem/machinetracker/src/MachineTrackerTemplate.tmpl
  Added note about inactive checkbox being useless for IPv6 searches.