Bugfix for LP#303120 (Machine Tracker switch search by IP address produces erroneous results)
This fixes and expands on the issues described in LP#303120.  Any
string entered in the "Switch" field in the search form is now the
subject of DNS reverse lookups and netbox table lookups.

When the user enters a valid IP address, a reverse DNS lookup is
performed.  Both the IP address and name are used as cam search
criteria, but both are also used to look up matching netbox records.
Any matching netbox records are used to expand further on the cam
search criteria.

One upshot of this is that historic search results will also reveal
records registered under a previous name the switch may have had
(which is seen a lot by some that have unstable DNS services).

* subsystem/machinetracker/src/machinetracker.py
  (hostname): Remove view logic, i.e. return None for failed lookups
  instead of the string "--".
  (ResultRow.__init__): Translate failed name lookups into the "--"
  string for display.
  (SwPortSQLQuery.__init__): Delegate construction of search criteria
  to new static method.
  (SwPortSQLQuery.expand_criteria): New static method.
  (SwPortSQLQuery.netbox_lookup): New static method.