Add index to cam table to speed up a slow ipdevinfo query.
When ipdevinfo queries for the first cam record from a given netboxid, the
query issued by Django sorts the results by the start_time field and limits the
result to one row.  Unfortunately, the limit causes PostgreSQL's planner to
scan the start_time index, believing this to be the most efficient plan.

If the device the user is browsing is a server, it has no cam records at all,
but the planner fails to realize that.  I've seen upwards of 40-60 seconds used
waiting for this single query on a database with over 10 million cam rows.  Now
the planner will choose to use this index instead, which is several orders of
magnitude faster.

* doc/sql/indexes.sql, doc/sql/upgrades/trunk.sql
  (cam_netboxid_start_time_btree): New index.