Fix for LP#338655 (DNS checker doesn't report version)
This comments back in the server.bind chaos query to get the server
version. The query fails Microsoft DNS: first time it returns an empty
answers list, the following times it times out. I've added an exception
handler for the timeout, which is now silently ignored.

Version checking now works for bind servers, but some more testing
is needed with non-bind servers.