Fix for LP#395185 (Adding new components to a existing maintenance task forces you to create a new task)
Maintenance form was hardcoded to post to the "new view".  In
addition, the tool is somewhat confused about whether to use 'id' from
the GET arguments or 'edit_taskid' from the POST arguments to identify
the maintenance task in question.  This system looks like it is in
serious need of refactoring.

* subsystem/maintenance/nav/web/templates/MaintenanceNewTemplate.tmpl
  Make the form action content a template variable.

* subsystem/maintenance/nav/web/maintenance/handler.py
  (handler): Set page.action to either "new" or "edit?id=taskid"
  depending on what view the user selected.