ipdevpoll/plugins/modules: Complete first useable version of the plugin.
Discover and store all field replaceable modules with non-empty
serials.  Also look at mappings between physical ports and ifindexes
and estimate on which registered module an interface belongs.

Unable to do a complete test of this yet, because of problems with the
storage system.

TODO: Add ability to store submodule relationships.