getBoksMacs and networkDiscovery: Try to fix swportallowedvlan.hexstring parsing code.
ipdevpoll will ATM always pull this value from CISCO-VTP-MIB, meaning
that the most significant bit in the string refers to VLAN 0.  It will
also make sure the hexstring is at least 256 digits long, as parts of
NAV contains redundant code that will parse shorter or longer
hexstrings thinking that the least significant bit is VLAN 0.

This patch makes sure getBoksMacs and networkDiscovery parse any
hexstring of 256 or more digits as a MSB=VLAN 0 hexstring.  It also
tries to remove the ridiculous hardcoded maximum of 1023 VLANs (by
setting it to another ridiculous, but more true value of 4095).