ipdevpoll: Set shadow class primary keys during Django ORM interaction.
When a shadow object is asked to locate an existing model object in
the database, it will set its own primary key attribute if successful.

When a new object is saved by a JobHandler, the resulting primary key
will be set on the corresponding shadow object.  This is an attempt to
achieve consistency, as other shadow objects that haven't been saved
yet may refer to that very shadow instance.

* subsystem/ipdevpoll/nav/ipdevpoll/storage.py
  (Storage.get_primary_key_attribute,
   Storage.get_primary_key,
   Storage.set_primary_key): Various new methods to make working with
   primary keys easier.

  (Storage.get_existing_model): Use new primary key methods.  Upon
  successful retrieval of existing model object, set our own primary
  key to that of the retrieved object.

* subsystem/ipdevpoll/nav/ipdevpoll/schedule.py
  (JobHandler.perform_save): If the shadow object we just saved
  doesn't know its own primary key, we retrieve it from the newly
  saved Django model object.