seeddb: Manage ORM and pure SQL mix magic during bulk imports.
The code uses the same attribute name to refer to db columns and
object attributes, but the Django models' attributes to not map
directly to db column names.

This is just a patch to make the the existing madness work, by mapping
db column names to django model attribute names before creating a
Django queryset.

* subsystem/seeddb/src/seeddb.py
  (bulkImportParse): Map db columns to attribute names before checking
  for row uniqueness using Django.