build: Change the way version numbers are obtained, and add simple autogen.sh bootstrap.
A build version number must now be obtained using the version.sh script, which
will generate a version.m4 file, included in the configure script built to
automake/autoconf.  The version number is generated by the Mercurial Describe
extensions, bundled in the tools directory.

For now, this is a manual process, but should probably be part of some sort of
post-commit-hook.