ipdevpoll: Handle binary garbage, as SNMP-collected from a catalyst6509.
Though there could be more cases of this in the code, a tested Catalyst 6509
running on IOS 12.2(33)SXI2a responded with non-UTF8-encodeable (binary)
strings as values for the model name and hardware revision for a GBic module
when the ENTITY-MIB was queried.

Symptoms of this are PostgreSQL errors about invalid UTF-8 strings.

* subsystem/ipdevpoll/nav/ipdevpoll/utils.py
  (is_invalid_utf8): New function.

* subsystem/ipdevpoll/nav/ipdevpoll/tests.py
  Added tests for new utils function.

* subsystem/ipdevpoll/nav/ipdevpoll/shadows.py
  (Module, Device): Check for "binary" version, serial and model name strings
  and convert them to their Python representations for DB storage.