Add check to see database connection status in servicemon
Servicemon assumes an empty list of services to check is caused by a dead
connection to the database, resulting in removed services to still be checked.

If the list is empty, and the connection is alive, flush this list.

Fixes LP#597569
(transplanted from 2f61e14e1a18565999975d614fd96609233ceba1)