avoid automake bug causing multiple recompilation of python modules during install.
For some reason, the PYTHON target causes autoconf to run a single instance of
the install binary for each subdirectory of the module hierarchy.  However,
the entire target directory is byte-compiled after each completed
subdirectory, causing the install process to take an inordinate amount of
time.

We don't really need these byte-compiled versions of the modules, so this
patch switches from using the PYTHON target to the DATA target when installing
python code into the pythondir.