(InterfaceManager.prepare): interfaces containers must be prepared _before_ loading existing objects.
this fixes a bug which caused interface updates to be interpreted as new
interfaces, because the interfaces plugin did not collect anything, but other
plugins had updates to interface vlan information. The Interface.prepare()
method would set an ifindex value from the container repo if it was missing,
but since this was done _after_ loading and matching existing objects, the
found objects would not be matched properly against the existing ones,
resulting in a database integrity violation as ipdevpoll tries to insert
interface records with duplicate ifindex values.