Fixed build_layer3_graph topology.
Layer3 topology isn't a directional graph, making it a MultiGraph. (Undirected graphs with self loops and parallel edges)
It's wrong to use the gwportprefix's interface.to_interface.netbox to get the netbox, as it gives you physical netbox it is connected to and not the logical one on the given prefix.
We need to build the graph using it's prefix as unique key when building the edges between netboxes. NAV's layer3 topology graph is based on GwRouterPrefixes as nodes,
later in nav.netmap.topology we will run thru graph and use Netboxes as node identifiers and add GwRouterPrefix-metadata to each related edge.