(is_netbox_reachable): improve algorithm in two steps.
if the router is up but not present in the graph, something is wrong
with the topology data. We default to saying the box should be reachable,
since we cannot properly determine its vlan topology.

(get_graph_for_vlan): put all nodes in graph.

(strip_down_nodes_from_graph): new function.