(vagrant-provision.sh) Updated provision script
Now it fetches build-dep for python-imaging so it can build PIL
Create database user and empty database for nav if not available.

Do some SEDing to enable django debug and set a default password
Set --prefix so we can run make in etc and python to generated
files that requires configured to be run. (ie buildconf.py)

At the end, run syncdb so we have all SQL changesets installed.