(servicemon) Remove unused code from servicemon daemon which depended on nav.path.
It was using nav.path.webroot which is deprecated module and 'webroot' variable. Since we're going to remove *.webroot usage as it is deprecated vs nav.buildconf.webrootdir, I abused my developer powers to remove these unused rotten code bits as it changelog is from 2003/2004 and this functionality is not in use at all anylonger and I se no reason to migrate unused code to new nav.buildconf.webrootdir usage.

IF THIS FUNCTIONALITY WAS IN USE, I guess we will hear it from our users and they should file a bug report requesting a feature which exports these recorded values to a file under webroot... WAAT for hardcoded commented code..