Remove accountproperties and use preferences instead.
AccountProperty was a separate table for storing preferences. This change
removes the AccountProperty model and changes all code from using that model to
using the hstore field 'preferences' on the Account model.

This affects only three things:
- The language setting in AlertProfiles (determines the language for alerts)
- The number of columns for widgets
- The status preferences for status-widgets and status page.

The code for personalizing the tool-setup also used AccountProperties. This code
was removed as it was not in use.