Move most logic into the frontend.
The backend now mostly prepares data, instead of doing complicated view stuff.
This solves the issue of tight coupling (and the eventual desync) of frontend
concerns with the backend (e.g. no need to generate strange-looking global IDs
for each DOM element we want to process or attach some event handler to).