Default all new Netboxes to SNMP v2c if version is omitted.
This is mostly an issue for bulk imported netboxes, where SNMP version is not
part of the import format.

v1-only agents will be reported as snmpAgentDown by NAV, mitigating this choice.
v1 agents that give the impression of being v2c capable but aren't are SOL,
but will result in other detectable errors.