[netmap] Land initial API change in frontend, add cache invalidation
This should improve performance by a lot, as we now only fetch traffic data for
the locations currently defined by our view. This data is cached as well, and
flushed either by a static timeout (on the backend) or by a flag ("?invalidate")
when calling the API. The refresh timer in the frontend has been rewritten to
use this flag, and the frontend itself to make use of the new API.

Stills needs work on optimalizing the layer 2 traffic query, which is painfully
slow. There also needs to be done some work providing a "sensible" default for a
new Netmap setup, for example an empty view (versus the current convention of
fetching everything, which is hilariously slow) and a wizard to set up a default
view to be used by the Netmap.

We also need to handle the case of multiple users triggering invalidation of the
traffic cache (L2/L3), as this might lead to poor performance due to interleaved
refresh intervals. I see no obvious solution to this problem except a ticket
system or other similarly complex (and over-engineered) coordination facilities,
so having a refresh rate might have been a bad idea to begin with :(

Signed-off-by: Emil Henry <emilhf@stud.ntnu.no>