[netmap] Simplify cache invalidation (by avoiding it)
The previous implementation allowed the (API) user to invalidate the L2/L3
traffic cache by setting an invalidation flag. This doesn't really make any
sense, as pointed out by @lunkwill42, as NAV only aggregates traffic every five
minutes. Hence, the timeout is set to this interval on the backend. This avoids
the issue of interleaved cache invalidations due to multiple Netmap users firing
of invalidation requests.

The next goal is probably to move this invalidation more into the core of NAV.
This also allows building an index for such data which might be useful for other
parts of NAV, for example the Geomap, saving us from performing a rather
expensive operation when we really don't need to. (Cache is somewhat cheap as
well, and the aggregated data is well within what Memcached or even file-based
caches should be able to handle, roughly ~60-100 kB).

Signed-off-by: Emil Henry <emilhf@stud.ntnu.no>