(StatPorts): remove redundant collection in cases of virtualized instances
In the case of a master netbox, duplicate all collected port metrics
across all instance netboxes when shipping to Graphite.

In the case of a virtualized instance, just debug log a list of local
ports (that do not appear on the master) and exit.

There is _no_ collection of data from local ports at the moment, as this
would require a complete restructuring of things (i.e collection from
individual ports, rather than full column snmpwalks for the values we
want)