Fix issue of missing meta-properties (fixes #1528)
After the new multiprocess mode was committed, Netbox objects are no
longer passed from the dataloader into JobHandler instances, only netbox
IDs, which are used to load a new Netbox instance from the db.

The meta-properties `snmp_up` and `last_updated` are therefore lost. The
default implementation of `Plugin.can_handle()` will see the `snmp_up`
property as a None-value, thinking that the Netbox currently has SNMP
issues and thereby declining to handle it at the moment.

This causes any SNMP-dependent job to not run - without it even being
flagged as an error anywhere, since this is the intended behavior when a
device isn't responding to SNMP.