rework docker image for Python 3
python:3.5 is based off Debian Jessie, which doesn't provide a mod_wsgi
that is compatible with Python 3.5, so this image is rebased to use
debian:stretch (which comes with Python 3.5), with the necessary
adjustments.

NodeJS/NPM isn't provided by Debian Stretch, so using the install
recipes found online. The integration tests don't complete under Python
3 yet, so we're not sure the JavaScript tests will work. Further testing
on Python 2 needs to be done first.