fix netmap bug that only appears under Pythons with random hash seeds
This code only worked because of consistent ordering of nodes in edges
under Python 2, but the tests would fail randomly under Python 3.

Adding the `id` attribute to `stubs.Netbox` fixes the tests, but one
test also assumes directed edges in an undirected graph (that the test
case builds itself!), so this was also fixed.