use edge metadata to describe edge when building JSON representations
the (u,v) coming from graph iteration may not always match up with
the (u,v) as described by the associated Edge objects, but (v,u)
instead.

Since the output is dealing with the notion of a source->target
direction, they should at least be in the same order in both in the
links list and the associated metadata.