DO NOT update the session cookie on every request
This would create some sort of race condition with parallel requests, that
causes the user to be inadvertently logged out just after logging in. Actual
users will experience this, and it would also cause intermittent failures of
the selenium tests (where the screenshot show the NAV login page instead of the
page under test).

We need to find a different strategy to update the session expiry time based on
user activity - AND this commit needs to be cherry picked onto the latest
stable branch once this strategy is in place.