Set audit log table width to 100%
DataTables overrides the table width if not set explicitly by a width
attribute. Setting it by use of a class does not work.

As the width seems to be set to 0px when the table is initially
hidden (typically in an expandable element) or have no data, this leads to
display problems as it needs to be explicitly recalculated when the element is
shown.