refactor report to use parametrized SQL queries (fixes #1711)
finally found a way to rework report to do it the "proper way"(tm), instead of
relying on pre-quoting of values using `nav.db.escape` or the psycopg2
extensions api.

ReportConfig now has a `parameters` list attribute, which must be appended to
for each filter expression that is added to a report.