Return 404 when asked for traffic from non-existant locations
Outdated Netmap views may have filters on rooms or locations that no longer
exist, which would cause a traceback to be mailed to the admin every time
such a map is populated with traffic data.

The correct and semantic way is not to crash, but to return an HTTP 404 status.