Switch the Docker development image to Python 3!
Our test coverage is still not good enough to catch all Python 3 problems, but
switching the dev environment over to Python 3 is a good start, as it will
bring to light Python 3 issues during development. Our goal is to have the next
feature release run entirely on Python 3.

Supervisor still requires Python 2 to run, unfortunately, so we still need it
in the base image, even though NAV itself doesn't have to touch it.

This also cleans up obsolete remnants from the environment (such as Apache,
which has not been used in the dev image for a long time).