Ensure last_used filter works as expectd on interface api endpoint
The InterfaceWithCamSerializer did not properly specify which model it was
using. Maybe older versions of REST framework would assume this from the
inherited class' Meta, but newer versions do not.

Also, there was an inherent crash bug with the last_used serializer, which
would crash whenever attempting to fetch a cam for for an interface that didn't
have any (which would be all the time if just fetching
/api/1/interfaces/?last_used=on)

Fixes #1839