Fix IPAM subnet allocation suggestions for ipv6
Changes UI to ask for a prefix length, rather than number of addresses wanted. Assume NAV users are networking people and know how this works. Asking for number of addresses makes no sense for ipv6.
The partition_subnet() function is essensially the same as the ipaddress modules IPXNetwork.subnets() function, so use that (include polyfill on python2)

Fixes #1594