Define SQL schema for new management_profile table.
This table will be used in combination with the netbox_profile table to define
management profiles that can be attached to zero or more netboxes.

The changescript also creates initial profiles from the existing netboxes,
removes the old SNMP fields from Netbox and attaches the corresponding profiles
to the existing netboxes.