Import reverse() from temporary location
This gets rid of the warning `RemovedInDjango20Warning: Importing from
django.core.urlresolvers is deprecated in favor of django.urls.`

This was a sed-replace. When on Django 1.11 or newer, sed it back like
so:

   sed -i 's/from nav.six import reverse/from django.urls import reverse/'

Because it was just sedded, pylint is unhappy because there's an import
from nav.* among django imports, so you get an import order warning.
Therefore, temporarily raise the pylint quality threshold:

Old: 1660
New: 1840

When on Django 1.11+ and after sedding to the modern location
(django.urls), remember to set the quality thresholds back to their old
values.