Rename all smidump modules to match MIB module names.
Since we now have nav.smidumps.get_mib to perform the module loads
dynamically, it doesn't matter that the module names aren't valid Python
names.

This way, get_mib() does not need to preload every module to discover
which MIB module it corresponds to. It can simply import modules by name
when needed.

It also makes it very easy to search the NAV code for the MIBs.