Implement dynamic form switching for mgmt protocol specific data
This enabled the SeedDB interface to edit management profiles to
dynamically switch out a fieldset of form fields related to specific
protocol implementations. By selecting a protocol from the protocol
dropdown, the correct fieldset for that protocol is made visible, while
all the other protocol-specific fieldsets are hidden (and their form
fields disabled).