Filter out UserWarnings from the psycopg2 module.
psycopg2 keeps issuing warnings about renaming of the binary
package. This is wholly irrelevant to NAV programs, and very annoying
for users.

The NAV development container is also set up to build the psycopg2
package from source, and that PyPi started providing binaries at a later
date was only incidental and inconsequential to NAV.