Apply "clean unicode" step for every use of EntityTable.
The "clean unicode" step is now mostly taken care of during the
collection phase, before the data hits an EntityTable instance. However,
objects that aren't marked up as displaystrings in the MIB will stay
bytes objects under Python 3. This would break this code under Python 3
whenever the "undefined datetime" string appears in the
ENTITY-MIB::entPhysicalMfgDate column, since this would still be a bytes
object as the parser attempts to parse it - and consequently ignores it
as non-string data. The raw string of null bytes is then later attempted
inserted into the database as a timestamp, which fails miserably.