Trust CAM data topology when there is no data from neighbor.
The rewritten topology code from 2017 lost one important piece: Servers
will mostly not be SNMP-enabled, and so we will have no candidate data
from servers. If there is one, and only one recognized CAM record on a
port, but no data at all from the other end, this patch will opt to
trust the data from that port (as it did, pre-2017), under the
assumption that the other end is unable to provide neighboring
information.

Fixes #1753