Use juniperdot1q as the default dot1q plugin
More and more people use NAV with Juniper switches. The presence of the
juniperdot1 plugin, and the option to use it instead of the regular
dot1q plugin in Juniper shops is not very obvious to the user. Since the
plugin does not break functionality for non-Juniper-devices, this
example config change will apply it as the default implementation for
dot1q in ipdevpoll.conf - which should be more user friendly to Juniper
users.