Generate traffic metric names rather than use graphite requests
For a function that purports to use a single request to get traffic
data, it seems kind of odd that it would make a graphite-web query to
discover the traffic counters of each and every interface.

Instead, just generate the known names of the wanted counters, and ask
for them. If they're not there, we just won't get a response.