Refactor ipdevpoll typeoid plugin
The plugin had a bug in which a Netbox would not have its type properly
set on the first run if a new type was created. The result of which
would be that when a new device was added to NAV, its type would not be
properly set for 6 hours (until the next inventory run). Lots of NAV
code paths are selected based on the registered device vendor, and so
the wrong code-paths would be chosen for the new device for these 6
hours, resulting in loss of proper monitoring.

The code of the plugin was hard to follow, as it was written before
inlineCallbacks were a thing, so I modernized the entire thing to use
said inlineCallbacks model.

To fix the issues of the proper creation and assignment of a new type
(and vendor), I changed the code to bypass the storage framework of
ipdevpoll and insert the new type/vendor directly into the database in a
db worker thread.