Consider AlertGenerator True on both PY2 and PY3
The __nonzero__() implementation was rendered useless under Python 3,
which replaced __nonzero__() with __bool__(). This in turn led to
AlertGenerator objects being considered False on Python 3, which caused
delayed posting of alerts using DelayedStateHandler to stop working.

Basically, no netboxes would ever be reported as down when running on
Python 3, which would have hade a huge impact on NAV. Luckily, there's a
black-box integration test for this issue now.