Restructure the installation documentation.
Because:
- The installation guide was too focused on installation from source
  code, and most users do not install NAV from source code.
- In case someone bypasses the initial information on
  downloading/installing NAV from the official homepage, the various
  options should be summarized in the official docs.

This changes the install doc to become a summary of installation
options, with links to the guides for each option. The generic
from-source guide is moved to a separate document. The doc on how to
integrate Graphite and NAV was also moved to a separate document (so it
can be referenced in multiple places), as this step is necessary for
most install options except for the virtual appliance or the docker
compose-based installation.