Debug log job dispatch event from master
Because:

- It's useful to discover a situation where the master process
  dispatched a job to a worker, but the worker never received the
  job. Therefore, this logs the dispatch event using more or less the same
  wording/syntax as the worker uses to log the reception of a job.

In addition, this adds the netboxid to the debug log output when a
JobHandler starts, to aid in debugging. The master dispatch/worker
reception code only works with netboxids (db primary keys), while the
JobHandler has received a complete Netbox object, but nowhere in the log
output was there a clear indication of the mapping between a logged
netboxid and a logged netbox sysname.