Prevent potential logic bug in syslog analyzer
Because:
- request.GET.keys() will always evaluate to True under Python 3,
since it returns an iterator/generator and not a list.

This fixes the same kind of problem as
127e3b57366a5ba217ec6963c60ada1e706186f1 did, but it's not clear that
anything was fundamentally broken in this instance, as the state ends up
being mostly the same in both code branches, since the form will not
turn up as "valid" when there is no actual data to be processed.