Remove superfluous get_native_vlan() method
Because:
- The method was never used outside of a ManagementHandler class.
- Apparently, its purpose was the exact same as
  get_interface_native_vlan(), and all the SNMPHandler base class
  implementation ever did was call that method.
- The Dell DNOS specific handler overrode the method, but there was no
  apparent reason why it shouldn't just override the
  get_interface_native_vlan() method instead.

It seems the entire method was introduced as a means to support Dell
switches. It may be that the operation for setting the native vlan
differs on access ports and trunk ports on Dell switches, but we have no
good way of testing this without access to a Dell DNOS switch.