Edit existing alert messages in case of dupes
Because:
- If, for some reason, an AlertQueue or AlertHistory object already has
  a message that matches the uniqueness criteria of the one we are
  trying to post, then we should not hard fail, but optionally edit the
  existing message (though this code does not change existing
  messages).
- Hard fails are bad during event processing, as you might get into a
  situation where an alert is never closed.

This situation would normally never happen, as processing of each event
in the queue will run fully inside a transaction, for example preventing
the end messages from being persisted to the database if updating the
end_time of the alerthist entry fails. Yet, this seems to have happened
on at least one customer install - so this attemps to take precautions
against that situation occurring again.

Fixes #2224