Implement a SeverityRules class w/tests
This class will serve as the main immutable representation of a ordered
sequence of severity modification rules. It provides functionality to
evaluate the rules in the context of a given alert, and return a new
severity value for the alert based on those rules.

Most of the public interface of the severity module is concentrated into
this single class.