Make AlertGenerator capable of using SeverityRules
AlertGenerator instances are used in many places, but it is not a given
that it should load severity rules directly from a specified NAV config
file. Since instances are created in so many places, it's also not a
given that each instantiator should duplicate the loading of severity
rules.

Therefere, a client application can initialize a SeverityRules
instance and set the AlertGenerator.severity_rules class variable in its
own process space, thereby defining application-wide rules.