Switch to six from django.utils.six
In order for the tests to run at all, six must be working. Better to
do a global change that temporarily fixes the problem everywhere and
then get rid of six, then having to reanalyse all places that uses six
while the tests are not working.

As long as we support Django 2.2, which is the last to support Python
2.7, we need something like six. When we no longer support 2.2, we can
get rid of six *and all other workrounds for running on Python 2.7 and
Python 3 at the same time* in a "stop supporting python 2.7"-branch.