Update docker dev image to Debian Bullseye
We need this to work an all Python 3.9 related issues in dev, since
Python 3.9 is the default Python version on the latest Debian stable
distro.

NET-SNMP has jumped from approx 5.7 to 5.9 (libsnmp30 vs libsnmp40), and
python3-gammu is no longer available in Debian (but we can get it from
PyPI if needed).