Disable PyLint again
It was unintentionally enabled in
5159f997bde23a5608b01c8aba767913e5451e03

Due to NAV having a huge legacy codebase before PyLint was ever
introduced, we've only look at PyLint violations as advisory. I.e.,
we've considered our metric to be that the number of violations flagged
by PyLint should always be decreasing, not increasing.

However, superlinter will lint all changed files fully, and this causes
it to flag old PyLint violations that aren't even part of the
commit. Also, superlinter fails the entire build for any violation,
which is NOT what we want.

It isn't clear how or if we can achieve the old model we were using with
Jenkins, where a number of violations are considered ok, as long as they
aren't new ones.