Fix syntax warning (and likely broken logic)
Newer Python versions issue a SyntaxWarning when this code is being run
in the test suite. This code is old, and this logic is entirely broken,
as suggested by the SyntaxWarning. Not exactly sure if or how this has
affected production code, but it's been there for the last 8 years
without raising any eyebrows :-P