Add tests for issue with Rack._configuration
Rack._configuration was defined as a VarcharField in the Django models,
while the actual PostgreSQL definition was a JSONB field. This
apparently worked fine with older version of Django, where the field was
magically made into a dict somwhere inside either Django or Psycopg2.

However, under Django 3, this assumption doesn't seem to hold any
longer: Now, the value is returned as a str, yet
theRack.configuration() property tries to access it as a dict, which
causes a crash.