Make Rack._configuration a proper JSONField
Older versions of Django/psycopg2 could apparently translate a
PostgreSQL JSONB field into a dict, while the dict needed to be
converted to a string before being saved to the database (which properly
saved the string as JSON).

However, this doesn't work anymore under Django 3. JSONField is now a
built-in in Django, and should be used directly instead.