Set state instead of event_type
Seems to be a bug. It previosuly set event.event_type in a place
where it (probably) should set event.state instead.
Found this by getting an error while trying
to create a stateless event. Seems this notify method is not
called anywhere, so makes sense it hasnt been found