Give thresholds generic nameing
makes it so the graphite metric paths have generic names,
instead of names specific to the juniper columns.
Any sensor threshold will have the same naming,
no matter what type of sensor it is. all you need
to know is the netbox, the sensor and the type of
threshold you want (high alarm, low alarm etc)