Do not create .eggs owned by root
There was a race condition between containers on `docker-compose up`: If
the docbuild container wins the race, it leaves a directory `.eggs`,
owned by root` in the mounted source directory. This prevents all the
other NAV-derived containers from making changed to this directory, and
they potentially crash on startup.

Running `setup.py develop` as root inside the docbuild container caused
this. Running `setup.py build` as the `nav` user first ensures the .eggs
are created without root privileges. The `nav` user's UID is mapped by
the container to the UID that owns the mounted /source directory.