Move the snmp_agent_proxy fixture
The snmp_agent_proxy fixture will be needed for more integration tests
than just the ones in mibs/sensor_test.

However, because of the new placement in integration/conftest.py, the
necessary imports would disturb the configuration of pytest. Therefore,
the imports needed to move to inside the fixture - and the global ports
variable needed to become a session fixture.