Ensure NetboxType is loaded in a db thread
This would run synchronously in the MainThread, which is a big no-no.
It would block the entire MainThread while waiting for the ORM result
from PostgreSQL, and also potentially attempt to use closed database
connections withouth the connection re-establishment precautions that
come with the database thread.

In short: MainThread is not designed to access the database at all!

Fixes #2478