Update zoom correctly when requesting link to configuration
Before, zoom value was ignored in view, but was present in the url, when user requested a link to current configuration. This commit makes zoomToBound() apply changes to zoom, when parameters are present. Parameters are present if the url with zoom, lon, lat values is present, ie when user requests link to current configuration.