Stop creating entity devices for ignored serials
This makes the entity plugin re-use the config setting from the
modules plugin. We still collect entity objects for everything in
entPhysicalTable, but this will not associate a Device entry with that
Entity if the device serial number is in the list of ignored serials.

Otherwise, the problem of multiple netboxes relating to the "same"
device and overwriting hw/gw/sw revision will continue.